requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
python-dotenv==0.21.0
yarl==1.8.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cpu
torchaudio==0.12.1+cpu
torchvision==0.13.1+cpu